linuxdo/requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
Pillow==10.3.0
pytz==2024.1
Requests>=2.32.3
Telethon==1.35.0
